# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CPI Artifacts API
  version: 1.0.0
  description: API Telegram Mini App для сбора материалов мероприятий.
servers:
  - url: https://artifacts.62-113-105-225.sslip.io
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/admin/dashboard:
    get:
      tags: [admin]
      summary: Сводные показатели
      responses:
        '200': { description: Дашборд }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/admin/events:
    get:
      tags: [admin]
      summary: Все мероприятия
      responses:
        '200': { description: Список }
    post:
      tags: [admin]
      summary: Создать мероприятие
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventInput' }
      responses:
        '201': { description: Мероприятие создано }
  /api/v1/admin/events/{eventId}:
    get:
      tags: [admin]
      parameters:
        - { $ref: '#/components/parameters/EventId' }
      responses:
        '200': { description: Мероприятие }
    patch:
      tags: [admin]
      parameters:
        - { $ref: '#/components/parameters/EventId' }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EventInput' }
      responses:
        '200': { description: Мероприятие обновлено }
    delete:
      tags: [admin]
      parameters:
        - { $ref: '#/components/parameters/EventId' }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '204': { description: Мероприятие архивировано }
  /api/v1/admin/users:
    get:
      tags: [admin]
      summary: Участники с агрегатами
      parameters:
        - { name: eventId, in: query, schema: { type: string, format: uuid } }
        - { name: q, in: query, schema: { type: string } }
      responses:
        '200': { description: Страница участников }
...
  /api/v1/admin/artifacts/{artifactId}:
    patch:
      tags: [admin]
      summary: Изменить статус файла
      parameters:
        - { $ref: '#/components/parameters/ArtifactId' }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '200': { description: Файл обновлён }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cpi_artifacts_session
  parameters:
    CsrfToken:
      name: x-csrf-token
      in: header
      required: true
      schema: { type: string }
...
    EventId:
      name: eventId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ArtifactId:
      name: artifactId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    EventFormat:
      enum: [offline, online, hybrid]
    EventStatus:
      enum: [draft, published, running, finished, archived]
...
    EventInput:
      type: object
      required:
        - title
        - slug
        - shortCode
        - organizer
        - startsAt
        - endsAt
        - timezone
        - format
        - status
        - acceptUploadsFrom
        - acceptUploadsUntil
      properties:
        title: { type: string, maxLength: 300 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        shortCode: { type: string, maxLength: 24 }
        description: { type: [string, 'null'] }
        organizer: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        timezone: { type: string }
        venue: { type: [string, 'null'] }
        city: { type: [string, 'null'] }
        format: { $ref: '#/components/schemas/EventFormat' }
        status: { $ref: '#/components/schemas/EventStatus' }
        tags: { type: array, items: { type: string } }
        acceptUploadsFrom: { type: string, format: date-time }
        acceptUploadsUntil: { type: string, format: date-time }
        maxFileSizeBytes: { type: integer, minimum: 1 }
        allowedMimeTypes: { type: array, items: { type: string } }
        blockedExtensions: { type: array, items: { type: string } }
        directAccessEnabled: { type: boolean }
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: {}
  responses:
...
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...